Rename modules with explicit names
Reformatting included
