+++ floris/visualization.py
@@ -9,7 +9,7 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from floris.coordinate import Coordinate
+from .coordinate import Coordinate
 import matplotlib.pyplot as plt
 import numpy as np
